fix e2e test by adding values to usage statuses collection before the test run #230
