+++ test/e2e/conftest.py
@@ -3,6 +3,7 @@ Module providing test fixtures for the e2e tests.
 """
 
 from test.conftest import VALID_ACCESS_TOKEN
+from test.e2e.mock_schemas import USAGE_STATUSES
 
 import pytest
 from fastapi.testclient import TestClient
@@ -19,6 +20,17 @@ def fixture_test_client() -> TestClient:
     :return: The test client.
     """
     return TestClient(app, headers={"Authorization": f"Bearer {VALID_ACCESS_TOKEN}"})
+
+
+@pytest.fixture(name="add_usage_statuses_to_database", autouse=True, scope="session")
+def fixture_add_usage_statuses_to_database():
+    """
+    Fixture to add usage statuses to the usage_statuses collection in the database before any tests in this file run
+    """
+    usage_statuses_collection = get_database().usage_statuses
+    usage_statuses_collection.delete_many({})
+    usage_statuses_collection.insert_many(USAGE_STATUSES)
+    yield
 
 
 @pytest.fixture(name="cleanup_database_collections", autouse=True)
